+++ mvc/build.gradle
@@ -24,6 +24,7 @@ dependencies {
     implementation "org.apache.commons:commons-lang3:3.13.0"
     implementation "ch.qos.logback:logback-classic:1.2.12"
 
+    testImplementation project(":app")
     testImplementation "org.assertj:assertj-core:3.24.2"
     testImplementation "org.junit.jupiter:junit-jupiter-api:5.7.2"
     testImplementation "org.mockito:mockito-core:5.4.0"
