+++ mvc/src/test/java/webmvc/org/springframework/web/servlet/mvc/tobe/DispatcherServletTest.java
@@ -1,4 +1,4 @@
-package com.techcourse;
+package webmvc.org.springframework.web.servlet.mvc.tobe;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatCode;
